changed the way mathjax renders stuff
